Extract key map extensions to a separate file
